refactored copy button into CopyButton component and implemented across AssistantMessage
