+++ components/Dialogue/MessageList/MessageItem/AssistantMessage/CopyButton/index.jsx
@@ -1,0 +1,65 @@
+import React, { useState, useEffect } from "react";
+
+function CopyButton({ text, copyMessage }) {
+  const [copied, setCopied] = useState(false);
+
+  const handleCopyClick = () => {
+    setCopied(true);
+    navigator.clipboard.writeText(text);
+  };
+
+  useEffect(() => {
+    setInterval(() => {
+      setCopied(false);
+    }, 1000);
+  }, [copied === true]);
+  return (
+    <button
+      className="flex ml-auto gap-2 copy"
+      onClick={() => {
+        handleCopyClick();
+      }}
+    >
+      {copied ? (
+        <>
+          <svg
+            stroke="currentColor"
+            fill="none"
+            strokeWidth="2"
+            viewBox="0 0 24 24"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            className="h-4 w-4"
+            height="1em"
+            width="1em"
+            xmlns="http://www.w3.org/2000/svg"
+          >
+            <polyline points="20 6 9 17 4 12"></polyline>
+          </svg>
+          {copyMessage?.postCopy && copyMessage.postCopy}
+        </>
+      ) : (
+        <>
+          <svg
+            stroke="currentColor"
+            fill="none"
+            strokeWidth="2"
+            viewBox="0 0 24 24"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            className="h-4 w-4"
+            height="1em"
+            width="1em"
+            xmlns="http://www.w3.org/2000/svg"
+          >
+            <path d="M16 4h2a2 2 0 0 1 2 2v14a2 2 0 0 1-2 2H6a2 2 0 0 1-2-2V6a2 2 0 0 1 2-2h2"></path>
+            <rect x="8" y="2" width="8" height="4" rx="1" ry="1"></rect>
+          </svg>
+          {copyMessage?.preCopy && copyMessage.preCopy}
+        </>
+      )}
+    </button>
+  );
+}
+
+export default CopyButton;
